# openapi/sms-verification.yaml
openapi: 3.1.0
info:
  title: Sendchamp SMS & Verification API
  description: |
    Sendchamp Open API for SMS messaging and OTP verification.
    Authenticate with your dashboard access key via Bearer token.
    Base URL: https://api.sendchamp.com/api/v1
  version: 1.0.0
  contact:
    name: Sendchamp Support
    url: https://support.go54.com/ticket/create
  license:
    name: Proprietary
    url: https://www.sendchamp.com/terms-of-service
servers:
  - url: https://api.sendchamp.com/api/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
  /sms/send:
    post:
      operationId: sendSms
      tags: [SMS]
      summary: Send SMS
      description: |
        Send a single SMS or multiple SMS in one request.
        Requires an approved sender name and sufficient wallet balance.
        Use route `non_dnd` for transactional/OTP traffic in Nigeria.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendSmsRequest'
            example:
              to: "+2348012345678"
              message: "Your order #1234 has shipped."
              sender_name: "MyBrand"
              route: "non_dnd"
              country_code: "NG"
      responses:
        '200':
          description: SMS accepted for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Dashboard access key from my.sendchamp.com → Settings → API Keys
  schemas:
    SendSmsRequest:
      type: object
      required: [to, message, sender_name, route]
      properties:
        to:
          oneOf:
            - type: string
              description: Single recipient in E.164 format
            - type: array
              items:
                type: string
              description: Multiple recipients in E.164 format
        message:
          type: string
          minLength: 1
        sender_name:
          type: string
          description: Approved sender ID registered on your account
        route:
          type: string
          enum: [dnd, non_dnd, international]
          description: |
            dnd = DND-registered Nigerian promotional traffic
            non_dnd = transactional/OTP (default for alerts)
            international = destinations outside Nigeria
        country_code:
          type: string
          description: ISO country code (e.g. NG, KE, GH)
...
    ApiSuccessResponse:
      type: object
      properties:
        status_code:
          type: integer
        message:
          type: string
        data:
          type: object
    ApiErrorResponse:
      type: object
      properties:
        status_code:
          type: integer
        message:
          type: string
        data:
          nullable: true
  responses:
    BadRequest:
      description: Invalid request or insufficient balance
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Invalid or missing access key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'